Move QuickCheck dep out of library
